pyproject.toml
[tool.poetry]
name = "buster-service"
version = "0.1.0"
description = "Chatbot service for Mila"
authors = ["Mila"]
license = "MIT"
readme = "README.md"
packages = [{include = "buster_service"}]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
buster = {git = "https://github.com/jerpint/buster", rev = "v1.0.14"}
lxml = "^4.9.3"
ovld = "^0.3.2"
grizzlaxy = "^0.0.2"
markdown = "^3.4.4"

[tool.poetry.scripts]
buster-gradio = "buster_service.app.__main__:main"
buster-acquire-sphinx = "buster_service.acquire.sphinx:main"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.285"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 120

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
